Start the migration due to changing the philosophy of the API.
Before, we didn't split concerns well given we had an explicit
expiryTime to the cache interface. We invert the dependency here by
storing and returning the createdAt timestamp when records are added to
the cache.

We also had a tight coupling on always importing the sqlite cache, even
if you implement your own cache (i.e. redis, duckdb). Now, we split
implementations of the cache into their own module, which can be
imported if required.

Business logic to check if a request/result is okay to store/return has
been improved in transport.go as well.

Overall improvements, but have removed tests for easier redesigning.
Will (hopefully) add before the final release.